Add SDXL PermSpec
Also attempt to fix unmatched key error
