+++ modules/merging/merge_PermSpec_SDXL.py
@@ -1,0 +1,1855 @@
+def sdxl_permutation_spec() -> PermutationSpec:
+    conv = lambda name, p_in, p_out: {  # pylint: disable=unnecessary-lambda-assignment
+        f"{name}.weight": (
+            p_out,
+            p_in,
+        ),
+        f"{name}.bias": (p_out,),
+    }
+    norm = lambda name, p: {f"{name}.weight": (p,), f"{name}.bias": (p,)}  # pylint: disable=unnecessary-lambda-assignment
+    dense = (
+        lambda name, p_in, p_out, bias=True: {  # pylint: disable=unnecessary-lambda-assignment
+            f"{name}.weight": (p_out, p_in),
+            f"{name}.bias": (p_out,),
+        }
+        if bias
+        else {f"{name}.weight": (p_out, p_in)}
+    )
+    skip = lambda name, p_in, p_out: {  # pylint: disable=unnecessary-lambda-assignment
+        f"{name}": (
+            p_out,
+            p_in,
+            None,
+            None,
+        )
+    }
+
+    # Unet Res blocks
+    easyblock = lambda name, p_in, p_out: {  # pylint: disable=unnecessary-lambda-assignment
+        **norm(f"{name}.in_layers.0", p_in),
+        **conv(f"{name}.in_layers.2", p_in, f"P_{name}_inner"),
+        **dense(
+            f"{name}.emb_layers.1", f"P_{name}_inner2", f"P_{name}_inner3", bias=True
+        ),
+        **norm(f"{name}.out_layers.0", f"P_{name}_inner4"),
+        **conv(f"{name}.out_layers.3", f"P_{name}_inner4", p_out),
+    }
+
+    return permutation_spec_from_axes_to_perm(
+        {
+            # Skipped Layers
+            **skip("betas", None, None),
+            **skip("alphas_cumprod", None, None),
+            **skip("alphas_cumprod_prev", None, None),
+            **skip("sqrt_alphas_cumprod", None, None),
+            **skip("sqrt_one_minus_alphas_cumprod", None, None),
+            **skip("log_one_minus_alphas_cumprods", None, None),
+            **skip("sqrt_recip_alphas_cumprod", None, None),
+            **skip("sqrt_recipm1_alphas_cumprod", None, None),
+            **skip("posterior_variance", None, None),
+            **skip("posterior_log_variance_clipped", None, None),
+            **skip("posterior_mean_coef1", None, None),
+            **skip("posterior_mean_coef2", None, None),
+            **skip("log_one_minus_alphas_cumprod", None, None),
+            **skip("model_ema.decay", None, None),
+            **skip("model_ema.num_updates", None, None),
+            **skip("model.diffusion_model.label_emb.0.0", None, None),
+            **skip("model.diffusion_model.label_emb.0.2", None, None),
+            **skip("model.diffusion_model.time_embed.0", None, None),
+            **skip("model.diffusion_model.time_embed.2", None, None),
+            **skip("conditioner.embedders.0.transformer.text_model.embeddings.position_ids", None, None),
+            **skip("conditioner.embedders.1.model.logit_scale", None, None),
+            **skip("conditioner.embedders.1.model.positional_embedding", None, None),
+            **skip("conditioner.embedders.1.model.text_projection", None, None),
+            **conv("model.diffusion_model.input_blocks.0.0", "P_bg0", "P_bg1"),
+            **easyblock("model.diffusion_model.input_blocks.1.0", "P_bg2", "P_bg3"),
+            **easyblock("model.diffusion_model.input_blocks.2.0", "P_bg4", "P_bg5"),
+            **conv("model.diffusion_model.input_blocks.3.0.op", "P_bg6", "P_bg7"),
+            **easyblock("model.diffusion_model.input_blocks.4.0", "P_bg8", "P_bg9"),
+            **conv("model.diffusion_model.input_blocks.4.0.skip_connection", "P_bg10", "P_bg11"),
+            **norm("model.diffusion_model.input_blocks.4.1.norm", "P_bg12"),
+            **conv("model.diffusion_model.input_blocks.4.1.proj_in", "P_bg12", "P_bg13"),
+            **conv("model.diffusion_model.input_blocks.4.1.proj_out", "P_bg14", "P_bg15"),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn1.to_k", "P_bg16", "P_bg17", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn1.to_out.0", "P_bg16", "P_bg17", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn1.to_q", "P_bg16", "P_bg17", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn1.to_v", "P_bg16", "P_bg17", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn2.to_k", "P_bg18", "P_bg19", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn2.to_out.0", "P_bg20", "P_bg21", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn2.to_q", "P_bg20", "P_bg21", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.attn2.to_v", "P_bg18", "P_bg19", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.ff.net.0.proj", "P_bg22", "P_bg23", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.ff.net.2", "P_bg24", "P_bg25", bias=True),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.norm1", "P_bg26"),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.norm2", "P_bg26"),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.0.norm3", "P_bg26"),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn1.to_k", "P_bg27", "P_bg28", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn1.to_out.0", "P_bg27", "P_bg28", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn1.to_q", "P_bg27", "P_bg28", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn1.to_v", "P_bg27", "P_bg28", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn2.to_k", "P_bg29", "P_bg30", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn2.to_out.0", "P_bg31", "P_bg32", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn2.to_q", "P_bg31", "P_bg32", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.attn2.to_v", "P_bg33", "P_bg34", bias=False),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.ff.net.0.proj", "P_bg35", "P_bg36", bias=True),
+            **dense("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.ff.net.2", "P_bg37", "P_bg38", bias=True),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.norm1", "P_bg39"),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.norm2", "P_bg39"),
+            **norm("model.diffusion_model.input_blocks.4.1.transformer_blocks.1.norm3", "P_bg39"),
+            **easyblock("model.diffusion_model.input_blocks.5.0", "P_bg40", "P_bg41"),
+            **norm("model.diffusion_model.input_blocks.5.1.norm", "P_bg42"),
+            **conv("model.diffusion_model.input_blocks.5.1.proj_in", "P_bg43", "P_bg44"),
+            **conv("model.diffusion_model.input_blocks.5.1.proj_out", "P_bg45", "P_bg46"),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn1.to_k", "P_bg47", "P_bg48", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn1.to_out.0", "P_bg47", "P_bg48", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn1.to_q", "P_bg47", "P_bg48", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn1.to_v", "P_bg47", "P_bg48", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn2.to_k", "P_bg49", "P_bg50", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn2.to_out.0", "P_bg51", "P_bg52", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn2.to_q", "P_bg53", "P_bg54", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.attn2.to_v", "P_bg55", "P_bg56", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.ff.net.0.proj", "P_bg57", "P_bg58", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.ff.net.2", "P_bg59", "P_bg60", bias=True),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.norm1", "P_bg61"),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.norm2", "P_bg61"),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.0.norm3", "P_bg61"),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn1.to_k", "P_bg62", "P_bg63", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn1.to_out.0", "P_bg62", "P_bg63", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn1.to_q", "P_bg62", "P_bg63", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn1.to_v", "P_bg62", "P_bg63", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn2.to_k", "P_bg64", "P_bg65", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn2.to_out.0", "P_bg66", "P_bg67", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn2.to_q", "P_bg68", "P_bg69", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.attn2.to_v", "P_bg70", "P_bg71", bias=False),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.ff.net.0.proj", "P_bg72", "P_bg73", bias=True),
+            **dense("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.ff.net.2", "P_bg74", "P_bg75", bias=True),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.norm1", "P_bg76"),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.norm2", "P_bg76"),
+            **norm("model.diffusion_model.input_blocks.5.1.transformer_blocks.1.norm3", "P_bg76"),
+            **conv("model.diffusion_model.input_blocks.6.0.op", "P_bg77", "P_bg78"),
+            **easyblock("model.diffusion_model.input_blocks.7.0", "P_bg78", "P_bg79"),
+            **conv("model.diffusion_model.input_blocks.7.0.skip_connection", "P_bg80", "P_bg81"),
+            **norm("model.diffusion_model.input_blocks.7.1.norm", "P_bg82"),
+            **conv("model.diffusion_model.input_blocks.7.1.proj_in", "P_bg82", "P_bg83"),
+            **conv("model.diffusion_model.input_blocks.7.1.proj_out", "P_bg84", "P_bg85"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn1.to_k", "P_bg86", "P_bg87", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn1.to_out.0", "P_bg86", "P_bg87", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn1.to_q", "P_bg86", "P_bg87", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn1.to_v", "P_bg86", "P_bg87", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn2.to_k", "P_bg88", "P_bg89", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn2.to_out.0", "P_bg90", "P_bg91", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn2.to_q", "P_bg92", "P_bg93", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.attn2.to_v", "P_bg94", "P_bg95", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.ff.net.0.proj", "P_bg96", "P_bg97", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.ff.net.2", "P_bg98", "P_bg99", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.norm1", "P_bg100"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.norm2", "P_bg100"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.0.norm3", "P_bg100"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn1.to_k", "P_bg101", "P_bg102", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn1.to_out.0", "P_bg101", "P_bg102", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn1.to_q", "P_bg101", "P_bg102", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn1.to_v", "P_bg101", "P_bg102", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn2.to_k", "P_bg103", "P_bg104", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn2.to_out.0", "P_bg105", "P_bg106", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn2.to_q", "P_bg107", "P_bg108", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.attn2.to_v", "P_bg109", "P_bg110", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.ff.net.0.proj", "P_bg111", "P_bg112", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.ff.net.2", "P_bg113", "P_bg114", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.norm1", "P_bg115"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.norm2", "P_bg115"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.1.norm3", "P_bg115"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn1.to_k", "P_bg116", "P_bg117", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn1.to_out.0", "P_bg116", "P_bg117", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn1.to_q", "P_bg116", "P_bg117", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn1.to_v", "P_bg116", "P_bg117", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn2.to_k", "P_bg118", "P_bg119", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn2.to_out.0", "P_bg120", "P_bg121", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn2.to_q", "P_bg122", "P_bg123", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.attn2.to_v", "P_bg124", "P_bg125", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.ff.net.0.proj", "P_bg126", "P_bg127", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.ff.net.2", "P_bg128", "P_bg129", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.norm1", "P_bg130"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.norm2", "P_bg130"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.2.norm3", "P_bg130"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn1.to_k", "P_bg110", "P_bg111", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn1.to_out.0", "P_bg131", "P_bg132", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn1.to_q", "P_bg131", "P_bg132", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn1.to_v", "P_bg131", "P_bg132", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn2.to_k", "P_bg133", "P_bg134", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn2.to_out.0", "P_bg135", "P_bg136", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn2.to_q", "P_bg137", "P_bg138", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.attn2.to_v", "P_bg139", "P_bg140", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.ff.net.0.proj", "P_bg141", "P_bg142", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.ff.net.2", "P_bg143", "P_bg144", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.norm1", "P_bg145"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.norm2", "P_bg145"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.3.norm3", "P_bg145"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn1.to_k", "P_bg146", "P_bg147", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn1.to_out.0", "P_bg146", "P_bg147", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn1.to_q", "P_bg146", "P_bg147", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn1.to_v", "P_bg146", "P_bg147", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn2.to_k", "P_bg148", "P_bg149", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn2.to_out.0", "P_bg150", "P_bg151", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn2.to_q", "P_bg152", "P_bg153", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.attn2.to_v", "P_bg154", "P_bg155", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.ff.net.0.proj", "P_bg156", "P_bg157", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.ff.net.2", "P_bg158", "P_bg159", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.norm1", "P_bg160"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.norm2", "P_bg160"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.4.norm3", "P_bg160"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn1.to_k", "P_bg161", "P_bg162", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn1.to_out.0", "P_bg161", "P_bg162", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn1.to_q", "P_bg161", "P_bg162", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn1.to_v", "P_bg161", "P_bg162", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn2.to_k", "P_bg163", "P_bg164", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn2.to_out.0", "P_bg165", "P_bg166", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn2.to_q", "P_bg167", "P_bg168", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.attn2.to_v", "P_bg169", "P_bg170", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.ff.net.0.proj", "P_bg171", "P_bg172", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.ff.net.2", "P_bg173", "P_bg174", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.norm1", "P_bg175"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.norm2", "P_bg175"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.5.norm3", "P_bg175"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn1.to_k", "P_bg176", "P_bg177", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn1.to_out.0", "P_bg176", "P_bg177", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn1.to_q", "P_bg176", "P_bg177", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn1.to_v", "P_bg176", "P_bg177", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn2.to_k", "P_bg178", "P_bg179", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn2.to_out.0", "P_bg180", "P_bg181", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn2.to_q", "P_bg182", "P_bg183", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.attn2.to_v", "P_bg184", "P_bg185", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.ff.net.0.proj", "P_bg186", "P_bg187", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.ff.net.2", "P_bg188", "P_bg189", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.norm1", "P_bg190"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.norm2", "P_bg190"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.6.norm3", "P_bg190"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn1.to_k", "P_bg191", "P_bg192", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn1.to_out.0", "P_bg193", "P_bg194", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn1.to_q", "P_bg195", "P_bg196", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn1.to_v", "P_bg197", "P_bg198", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn2.to_k", "P_bg199", "P_bg200", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn2.to_out.0", "P_bg201", "P_bg202", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn2.to_q", "P_bg203", "P_bg204", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.attn2.to_v", "P_bg205", "P_bg206", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.ff.net.0.proj", "P_bg207", "P_bg208", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.ff.net.2", "P_bg209", "P_bg210", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.norm1", "P_bg211"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.norm2", "P_bg211"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.7.norm3", "P_bg211"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn1.to_k", "P_bg212", "P_bg213", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn1.to_out.0", "P_bg212", "P_bg213", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn1.to_q", "P_bg212", "P_bg213", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn1.to_v", "P_bg212", "P_bg213", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn2.to_k", "P_bg214", "P_bg215", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn2.to_out.0", "P_bg216", "P_bg217", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn2.to_q", "P_bg218", "P_bg219", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.attn2.to_v", "P_bg220", "P_bg221", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.ff.net.0.proj", "P_bg222", "P_bg223", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.ff.net.2", "P_bg224", "P_bg225", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.norm1", "P_bg226"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.norm2", "P_bg226"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.8.norm3", "P_bg226"),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn1.to_k", "P_bg227", "P_bg228", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn1.to_out.0", "P_bg227", "P_bg228", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn1.to_q", "P_bg227", "P_bg228", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn1.to_v", "P_bg227", "P_bg228", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn2.to_k", "P_bg229", "P_bg230", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn2.to_out.0", "P_bg231", "P_bg232", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn2.to_q", "P_bg233", "P_bg234", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.attn2.to_v", "P_bg235", "P_bg236", bias=False),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.ff.net.0.proj", "P_bg237", "P_bg238", bias=True),
+            **dense("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.ff.net.2", "P_bg239", "P_bg240", bias=True),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.norm1", "P_bg241"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.norm2", "P_bg241"),
+            **norm("model.diffusion_model.input_blocks.7.1.transformer_blocks.9.norm3", "P_bg241"),
+            **easyblock("model.diffusion_model.input_blocks.8.0", "P_bg242", "P_bg243"),
+            **norm("model.diffusion_model.input_blocks.8.1.norm", "P_bg244"),
+            **conv("model.diffusion_model.input_blocks.8.1.proj_in", "P_bg244", "P_bg245"),
+            **conv("model.diffusion_model.input_blocks.8.1.proj_out", "P_bg246", "P_bg247"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn1.to_k", "P_bg248", "P_bg249", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn1.to_out.0", "P_bg248", "P_bg249", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn1.to_q", "P_bg248", "P_bg249", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn1.to_v", "P_bg248", "P_bg249", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn2.to_k", "P_bg250", "P_bg251", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn2.to_out.0", "P_bg252", "P_bg253", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn2.to_q", "P_bg254", "P_bg255", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.attn2.to_v", "P_bg256", "P_bg257", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.ff.net.0.proj", "P_bg258", "P_bg259", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.ff.net.2", "P_bg260", "P_bg261", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.norm1", "P_bg262"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.norm2", "P_bg262"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.0.norm3", "P_bg262"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn1.to_k", "P_bg263", "P_bg264", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn1.to_out.0", "P_bg263", "P_bg264", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn1.to_q", "P_bg263", "P_bg264", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn1.to_v", "P_bg263", "P_bg264", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn2.to_k", "P_bg265", "P_bg266", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn2.to_out.0", "P_bg267", "P_bg268", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn2.to_q", "P_bg269", "P_bg270", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.attn2.to_v", "P_bg271", "P_bg272", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.ff.net.0.proj", "P_bg273", "P_bg274", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.ff.net.2", "P_bg275", "P_bg276", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.norm1", "P_bg277"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.norm2", "P_bg277"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.1.norm3", "P_bg277"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn1.to_k", "P_bg278", "P_bg279", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn1.to_out.0", "P_bg278", "P_bg279", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn1.to_q", "P_bg278", "P_bg279", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn1.to_v", "P_bg278", "P_bg279", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn2.to_k", "P_bg280", "P_bg281", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn2.to_out.0", "P_bg282", "P_bg283", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn2.to_q", "P_bg284", "P_bg285", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.attn2.to_v", "P_bg286", "P_bg287", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.ff.net.0.proj", "P_bg288", "P_bg289", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.ff.net.2", "P_bg290", "P_bg291", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.norm1", "P_bg292"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.norm2", "P_bg292"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.2.norm3", "P_bg292"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn1.to_k", "P_bg293", "P_bg294", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn1.to_out.0", "P_bg293", "P_bg294", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn1.to_q", "P_bg293", "P_bg294", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn1.to_v", "P_bg293", "P_bg294", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn2.to_k", "P_bg295", "P_bg296", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn2.to_out.0", "P_bg297", "P_bg298", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn2.to_q", "P_bg299", "P_bg300", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.attn2.to_v", "P_bg301", "P_bg302", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.ff.net.0.proj", "P_bg303", "P_bg304", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.ff.net.2", "P_bg305", "P_bg306", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.norm1", "P_bg307"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.norm2", "P_bg307"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.3.norm3", "P_bg307"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn1.to_k", "P_bg308", "P_bg309", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn1.to_out.0", "P_bg308", "P_bg309", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn1.to_q", "P_bg308", "P_bg309", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn1.to_v", "P_bg308", "P_bg309", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn2.to_k", "P_bg310", "P_bg311", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn2.to_out.0", "P_bg312", "P_bg313", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn2.to_q", "P_bg314", "P_bg315", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.attn2.to_v", "P_bg316", "P_bg317", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.ff.net.0.proj", "P_bg318", "P_bg319", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.ff.net.2", "P_bg320", "P_bg321", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.norm1", "P_bg322"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.norm2", "P_bg322"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.4.norm3", "P_bg322"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn1.to_k", "P_bg322", "P_bg323", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn1.to_out.0", "P_bg322", "P_bg323", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn1.to_q", "P_bg322", "P_bg323", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn1.to_v", "P_bg322", "P_bg323", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn2.to_k", "P_bg324", "P_bg325", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn2.to_out.0", "P_bg326", "P_bg327", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn2.to_q", "P_bg328", "P_bg329", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.attn2.to_v", "P_bg330", "P_bg331", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.ff.net.0.proj", "P_bg332", "P_bg333", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.ff.net.2", "P_bg334", "P_bg335", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.norm1", "P_bg336"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.norm2", "P_bg336"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.5.norm3", "P_bg336"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn1.to_k", "P_bg337", "P_bg338", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn1.to_out.0", "P_bg337", "P_bg338", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn1.to_q", "P_bg337", "P_bg338", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn1.to_v", "P_bg337", "P_bg338", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn2.to_k", "P_bg339", "P_bg340", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn2.to_out.0", "P_bg341", "P_bg342", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn2.to_q", "P_bg343", "P_bg344", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.attn2.to_v", "P_bg345", "P_bg346", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.ff.net.0.proj", "P_bg347", "P_bg348", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.ff.net.2", "P_bg349", "P_bg350", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.norm1", "P_bg351"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.norm2", "P_bg351"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.6.norm3", "P_bg351"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn1.to_k", "P_bg352", "P_bg353", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn1.to_out.0", "P_bg352", "P_bg353", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn1.to_q", "P_bg352", "P_bg353", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn1.to_v", "P_bg352", "P_bg353", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn2.to_k", "P_bg354", "P_bg355", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn2.to_out.0", "P_bg356", "P_bg357", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn2.to_q", "P_bg358", "P_bg359", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.attn2.to_v", "P_bg360", "P_bg361", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.ff.net.0.proj", "P_bg362", "P_bg363", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.ff.net.2", "P_bg364", "P_bg365", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.norm1", "P_bg366"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.norm2", "P_bg366"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.7.norm3", "P_bg366"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn1.to_k", "P_bg367", "P_bg368", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn1.to_out.0", "P_bg367", "P_bg368", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn1.to_q", "P_bg367", "P_bg368", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn1.to_v", "P_bg367", "P_bg368", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn2.to_k", "P_bg369", "P_bg370", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn2.to_out.0", "P_bg371", "P_bg372", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn2.to_q", "P_bg373", "P_bg374", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.attn2.to_v", "P_bg375", "P_bg376", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.ff.net.0.proj", "P_bg377", "P_bg378", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.ff.net.2", "P_bg379", "P_bg380", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.norm1", "P_bg381"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.norm2", "P_bg381"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.8.norm3", "P_bg381"),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn1.to_k", "P_bg382", "P_bg383", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn1.to_out.0", "P_bg382", "P_bg383", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn1.to_q", "P_bg382", "P_bg383", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn1.to_v", "P_bg382", "P_bg383", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn2.to_k", "P_bg384", "P_bg385", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn2.to_out.0", "P_bg386", "P_bg387", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn2.to_q", "P_bg388", "P_bg389", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.attn2.to_v", "P_bg390", "P_bg391", bias=False),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.ff.net.0.proj", "P_bg392", "P_bg393", bias=True),
+            **dense("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.ff.net.2", "P_bg394", "P_bg395", bias=True),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.norm1", "P_bg396"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.norm2", "P_bg396"),
+            **norm("model.diffusion_model.input_blocks.8.1.transformer_blocks.9.norm3", "P_bg396"),
+            **easyblock("model.diffusion_model.middle_block.0", "P_bg397", "P_bg398"),
+            **norm("model.diffusion_model.middle_block.1.norm", "P_bg399"),
+            **conv("model.diffusion_model.middle_block.1.proj_in", "P_bg399", "P_bg400"),
+            **conv("model.diffusion_model.middle_block.1.proj_out", "P_bg401", "P_bg402"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn1.to_k", "P_bg403", "P_bg404", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn1.to_out.0", "P_bg403", "P_bg404", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn1.to_q", "P_bg403", "P_bg404", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn1.to_v", "P_bg403", "P_bg404", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn2.to_k", "P_bg405", "P_bg406", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn2.to_out.0", "P_bg407", "P_bg408", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn2.to_q", "P_bg409", "P_bg410", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.attn2.to_v", "P_bg411", "P_bg412", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.ff.net.0.proj", "P_bg413", "P_bg414", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.0.ff.net.2", "P_bg415", "P_bg416", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.0.norm1", "P_bg417"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.0.norm2", "P_bg417"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.0.norm3", "P_bg417"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn1.to_k", "P_bg418", "P_bg419", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn1.to_out.0", "P_bg418", "P_bg419", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn1.to_q", "P_bg418", "P_bg419", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn1.to_v", "P_bg418", "P_bg419", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn2.to_k", "P_bg420", "P_bg421", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn2.to_out.0", "P_bg422", "P_bg423", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn2.to_q", "P_bg424", "P_bg425", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.attn2.to_v", "P_bg426", "P_bg427", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.ff.net.0.proj", "P_bg428", "P_bg429", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.1.ff.net.2", "P_bg430", "P_bg431", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.1.norm1", "P_bg432"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.1.norm2", "P_bg432"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.1.norm3", "P_bg432"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn1.to_k", "P_bg433", "P_bg434", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn1.to_out.0", "P_bg433", "P_bg434", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn1.to_q", "P_bg433", "P_bg434", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn1.to_v", "P_bg433", "P_bg434", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn2.to_k", "P_bg435", "P_bg436", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn2.to_out.0", "P_bg437", "P_bg438", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn2.to_q", "P_bg439", "P_bg440", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.attn2.to_v", "P_bg441", "P_bg442", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.ff.net.0.proj", "P_bg443", "P_bg444", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.2.ff.net.2", "P_bg445", "P_bg446", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.2.norm1", "P_bg447"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.2.norm2", "P_bg447"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.2.norm3", "P_bg447"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn1.to_k", "P_bg448", "P_bg449", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn1.to_out.0", "P_bg448", "P_bg449", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn1.to_q", "P_bg448", "P_bg449", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn1.to_v", "P_bg448", "P_bg449", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn2.to_k", "P_bg450", "P_bg451", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn2.to_out.0", "P_bg452", "P_bg453", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn2.to_q", "P_bg454", "P_bg455", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.attn2.to_v", "P_bg456", "P_bg457", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.ff.net.0.proj", "P_bg458", "P_bg459", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.3.ff.net.2", "P_bg460", "P_bg461", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.3.norm1", "P_bg462"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.3.norm2", "P_bg462"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.3.norm3", "P_bg462"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn1.to_k", "P_bg463", "P_bg464", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn1.to_out.0", "P_bg463", "P_bg464", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn1.to_q", "P_bg463", "P_bg464", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn1.to_v", "P_bg463", "P_bg464", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn2.to_k", "P_bg465", "P_bg466", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn2.to_out.0", "P_bg467", "P_bg468", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn2.to_q", "P_bg469", "P_bg470", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.attn2.to_v", "P_bg471", "P_bg472", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.ff.net.0.proj", "P_bg473", "P_bg474", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.4.ff.net.2", "P_bg475", "P_bg476", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.4.norm1", "P_bg477"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.4.norm2", "P_bg477"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.4.norm3", "P_bg477"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn1.to_k", "P_bg478", "P_bg479", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn1.to_out.0", "P_bg478", "P_bg479", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn1.to_q", "P_bg478", "P_bg479", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn1.to_v", "P_bg478", "P_bg479", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn2.to_k", "P_bg480", "P_bg481", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn2.to_out.0", "P_bg482", "P_bg483", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn2.to_q", "P_bg484", "P_bg485", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.attn2.to_v", "P_bg486", "P_bg487", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.ff.net.0.proj", "P_bg488", "P_bg489", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.5.ff.net.2", "P_bg490", "P_bg491", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.5.norm1", "P_bg492"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.5.norm2", "P_bg492"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.5.norm3", "P_bg492"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn1.to_k", "P_bg493", "P_bg494", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn1.to_out.0", "P_bg493", "P_bg494", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn1.to_q", "P_bg493", "P_bg494", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn1.to_v", "P_bg493", "P_bg494", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn2.to_k", "P_bg495", "P_bg496", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn2.to_out.0", "P_bg497", "P_bg498", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn2.to_q", "P_bg499", "P_bg500", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.attn2.to_v", "P_bg501", "P_bg502", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.ff.net.0.proj", "P_bg503", "P_bg504", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.6.ff.net.2", "P_bg505", "P_bg506", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.6.norm1", "P_bg507"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.6.norm2", "P_bg507"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.6.norm3", "P_bg507"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn1.to_k", "P_bg508", "P_bg509", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn1.to_out.0", "P_bg508", "P_bg509", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn1.to_q", "P_bg508", "P_bg509", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn1.to_v", "P_bg508", "P_bg509", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn2.to_k", "P_bg510", "P_bg511", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn2.to_out.0", "P_bg512", "P_bg513", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn2.to_q", "P_bg514", "P_bg515", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.attn2.to_v", "P_bg516", "P_bg517", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.ff.net.0.proj", "P_bg518", "P_bg519", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.7.ff.net.2", "P_bg520", "P_bg521", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.7.norm1", "P_bg522"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.7.norm2", "P_bg522"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.7.norm3", "P_bg522"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn1.to_k", "P_bg523", "P_bg524", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn1.to_out.0", "P_bg523", "P_bg524", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn1.to_q", "P_bg523", "P_bg524", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn1.to_v", "P_bg523", "P_bg524", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn2.to_k", "P_bg525", "P_bg526", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn2.to_out.0", "P_bg527", "P_bg528", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn2.to_q", "P_bg529", "P_bg530", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.attn2.to_v", "P_bg531", "P_bg532", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.ff.net.0.proj", "P_bg533", "P_bg534", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.8.ff.net.2", "P_bg535", "P_bg536", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.8.norm1", "P_bg537"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.8.norm2", "P_bg537"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.8.norm3", "P_bg537"),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn1.to_k", "P_bg538", "P_bg539", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn1.to_out.0", "P_bg538", "P_bg539", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn1.to_q", "P_bg538", "P_bg539", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn1.to_v", "P_bg538", "P_bg539", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn2.to_k", "P_bg540", "P_bg541", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn2.to_out.0", "P_bg542", "P_bg543", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn2.to_q", "P_bg544", "P_bg545", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.attn2.to_v", "P_bg546", "P_bg547", bias=False),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.ff.net.0.proj", "P_bg548", "P_bg549", bias=True),
+            **dense("model.diffusion_model.middle_block.1.transformer_blocks.9.ff.net.2", "P_bg550", "P_bg551", bias=True),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.9.norm1", "P_bg552"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.9.norm2", "P_bg552"),
+            **norm("model.diffusion_model.middle_block.1.transformer_blocks.9.norm3", "P_bg552"),
+            **easyblock("model.diffusion_model.middle_block.2", "P_bg553", "P_bg554"),
+            **norm("model.diffusion_model.out.0", "P_bg555"),
+            **conv("model.diffusion_model.out.2", "P_bg556", "P_bg557"),
+            **easyblock("model.diffusion_model.output_blocks.0.0", "P_bg558", "P_bg559"),
+            **conv("model.diffusion_model.output_blocks.0.0.skip_connection", "P_bg560", "P_bg561"),
+            **norm("model.diffusion_model.output_blocks.0.1.norm", "P_bg562"),
+            **conv("model.diffusion_model.output_blocks.0.1.proj_in", "P_bg562", "P_bg563"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn1.to_k", "P_bg564", "P_bg565", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn1.to_out.0", "P_bg564", "P_bg565", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn1.to_q", "P_bg564", "P_bg565", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn1.to_v", "P_bg564", "P_bg565", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn2.to_k", "P_bg566", "P_bg567", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn2.to_out.0", "P_bg568", "P_bg569", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn2.to_q", "P_bg570", "P_bg571", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.attn2.to_v", "P_bg572", "P_bg573", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.ff.net.0.proj", "P_bg574", "P_bg575", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.ff.net.2", "P_bg576", "P_bg577", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.norm1", "P_bg578"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.norm2", "P_bg578"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.0.norm3", "P_bg578"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn1.to_k", "P_bg579", "P_bg580", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn1.to_out.0", "P_bg579", "P_bg580", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn1.to_q", "P_bg579", "P_bg580", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn1.to_v", "P_bg579", "P_bg580", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn2.to_k", "P_bg581", "P_bg582", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn2.to_out.0", "P_bg583", "P_bg584", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn2.to_q", "P_bg585", "P_bg586", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.attn2.to_v", "P_bg587", "P_bg588", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.ff.net.0.proj", "P_bg589", "P_bg590", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.ff.net.2", "P_bg591", "P_bg592", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.norm1", "P_bg593"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.norm2", "P_bg593"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.1.norm3", "P_bg593"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn1.to_k", "P_bg594", "P_bg595", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn1.to_out.0", "P_bg594", "P_bg595", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn1.to_q", "P_bg594", "P_bg595", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn1.to_v", "P_bg594", "P_bg595", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn2.to_k", "P_bg596", "P_bg597", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn2.to_out.0", "P_bg598", "P_bg599", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn2.to_q", "P_bg600", "P_bg601", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.attn2.to_v", "P_bg602", "P_bg603", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.ff.net.0.proj", "P_bg604", "P_bg605", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.ff.net.2", "P_bg606", "P_bg607", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.norm1", "P_bg608"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.norm2", "P_bg608"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.2.norm3", "P_bg608"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn1.to_k", "P_bg609", "P_bg610", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn1.to_out.0", "P_bg609", "P_bg610", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn1.to_q", "P_bg609", "P_bg610", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn1.to_v", "P_bg609", "P_bg610", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn2.to_k", "P_bg611", "P_bg612", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn2.to_out.0", "P_bg613", "P_bg614", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn2.to_q", "P_bg615", "P_bg616", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.attn2.to_v", "P_bg617", "P_bg618", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.ff.net.0.proj", "P_bg619", "P_bg620", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.ff.net.2", "P_bg621", "P_bg622", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.norm1", "P_bg623"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.norm2", "P_bg623"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.3.norm3", "P_bg623"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn1.to_k", "P_bg624", "P_bg625", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn1.to_out.0", "P_bg624", "P_bg625", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn1.to_q", "P_bg624", "P_bg625", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn1.to_v", "P_bg624", "P_bg625", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn2.to_k", "P_bg626", "P_bg627", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn2.to_out.0", "P_bg628", "P_bg629", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn2.to_q", "P_bg630", "P_bg631", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.attn2.to_v", "P_bg632", "P_bg633", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.ff.net.0.proj", "P_bg634", "P_bg635", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.ff.net.2", "P_bg636", "P_bg637", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.norm1", "P_bg638"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.norm2", "P_bg638"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.4.norm3", "P_bg638"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn1.to_k", "P_bg639", "P_bg640", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn1.to_out.0", "P_bg639", "P_bg640", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn1.to_q", "P_bg639", "P_bg640", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn1.to_v", "P_bg639", "P_bg640", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn2.to_k", "P_bg641", "P_bg642", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn2.to_out.0", "P_bg643", "P_bg644", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn2.to_q", "P_bg645", "P_bg646", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.attn2.to_v", "P_bg647", "P_bg648", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.ff.net.0.proj", "P_bg649", "P_bg650", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.ff.net.2", "P_bg651", "P_bg652", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.norm1", "P_bg653"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.norm2", "P_bg653"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.5.norm3", "P_bg653"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn1.to_k", "P_bg654", "P_bg655", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn1.to_out.0", "P_bg654", "P_bg655", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn1.to_q", "P_bg654", "P_bg655", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn1.to_v", "P_bg654", "P_bg655", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn2.to_k", "P_bg656", "P_bg657", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn2.to_out.0", "P_bg658", "P_bg659", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn2.to_q", "P_bg660", "P_bg661", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.attn2.to_v", "P_bg662", "P_bg663", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.ff.net.0.proj", "P_bg664", "P_bg665", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.ff.net.2", "P_bg666", "P_bg667", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.norm1", "P_bg668"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.norm2", "P_bg668"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.6.norm3", "P_bg668"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn1.to_k", "P_bg669", "P_bg670", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn1.to_out.0", "P_bg669", "P_bg670", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn1.to_q", "P_bg669", "P_bg670", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn1.to_v", "P_bg669", "P_bg670", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn2.to_k", "P_bg671", "P_bg672", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn2.to_out.0", "P_bg673", "P_bg674", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn2.to_q", "P_bg675", "P_bg676", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.attn2.to_v", "P_bg677", "P_bg678", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.ff.net.0.proj", "P_bg679", "P_bg680", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.ff.net.2", "P_bg681", "P_bg682", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.norm1", "P_bg683"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.norm2", "P_bg683"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.7.norm3", "P_bg683"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn1.to_k", "P_bg684", "P_bg685", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn1.to_out.0", "P_bg684", "P_bg685", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn1.to_q", "P_bg684", "P_bg685", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn1.to_v", "P_bg684", "P_bg685", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn2.to_k", "P_bg686", "P_bg687", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn2.to_out.0", "P_bg688", "P_bg689", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn2.to_q", "P_bg690", "P_bg691", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.attn2.to_v", "P_bg692", "P_bg693", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.ff.net.0.proj", "P_bg694", "P_bg695", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.ff.net.2", "P_bg696", "P_bg697", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.norm1", "P_bg698"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.norm2", "P_bg698"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.8.norm3", "P_bg698"),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn1.to_k", "P_bg699", "P_bg700", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn1.to_out.0", "P_bg699", "P_bg700", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn1.to_q", "P_bg699", "P_bg700", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn1.to_v", "P_bg699", "P_bg700", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn2.to_k", "P_bg701", "P_bg702", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn2.to_out.0", "P_bg703", "P_bg704", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn2.to_q", "P_bg705", "P_bg706", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.attn2.to_v", "P_bg707", "P_bg708", bias=False),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.ff.net.0.proj", "P_bg709", "P_bg710", bias=True),
+            **dense("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.ff.net.2", "P_bg711", "P_bg712", bias=True),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.norm1", "P_bg713"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.norm2", "P_bg713"),
+            **norm("model.diffusion_model.output_blocks.0.1.transformer_blocks.9.norm3", "P_bg713"),
+            **easyblock("model.diffusion_model.output_blocks.1.0", "P_bg714", "P_bg715"),
+            **conv("model.diffusion_model.output_blocks.1.0.skip_connection", "P_bg716", "P_bg717"),
+            **norm("model.diffusion_model.output_blocks.1.1.norm", "P_bg718"),
+            **conv("model.diffusion_model.output_blocks.1.1.proj_in", "P_bg718", "P_bg719"),
+            **conv("model.diffusion_model.output_blocks.1.1.proj_out", "P_bg720", "P_bg721"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn1.to_k", "P_bg722", "P_bg723", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn1.to_out.0", "P_bg722", "P_bg723", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn1.to_q", "P_bg722", "P_bg723", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn1.to_v", "P_bg722", "P_bg723", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn2.to_k", "P_bg724", "P_bg725", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn2.to_out.0", "P_bg726", "P_bg727", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn2.to_q", "P_bg728", "P_bg729", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.attn2.to_v", "P_bg730", "P_bg731", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.ff.net.0.proj", "P_bg732", "P_bg733", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.ff.net.2", "P_bg734", "P_bg735", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.norm1", "P_bg736"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.norm2", "P_bg736"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.0.norm3", "P_bg736"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn1.to_k", "P_bg737", "P_bg738", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn1.to_out.0", "P_bg737", "P_bg738", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn1.to_q", "P_bg737", "P_bg738", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn1.to_v", "P_bg737", "P_bg738", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn2.to_k", "P_bg739", "P_bg740", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn2.to_out.0", "P_bg741", "P_bg742", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn2.to_q", "P_bg743", "P_bg744", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.attn2.to_v", "P_bg745", "P_bg746", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.ff.net.0.proj", "P_bg747", "P_bg748", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.ff.net.2", "P_bg749", "P_bg750", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.norm1", "P_bg751"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.norm2", "P_bg751"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.1.norm3", "P_bg751"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn1.to_k", "P_bg752", "P_bg753", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn1.to_out.0", "P_bg752", "P_bg753", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn1.to_q", "P_bg752", "P_bg753", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn1.to_v", "P_bg752", "P_bg753", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn2.to_k", "P_bg754", "P_bg755", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn2.to_out.0", "P_bg756", "P_bg757", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn2.to_q", "P_bg758", "P_bg759", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.attn2.to_v", "P_bg760", "P_bg761", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.ff.net.0.proj", "P_bg762", "P_bg763", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.ff.net.2", "P_bg764", "P_bg765", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.norm1", "P_bg766"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.norm2", "P_bg766"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.2.norm3", "P_bg766"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn1.to_k", "P_bg767", "P_bg768", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn1.to_out.0", "P_bg767", "P_bg768", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn1.to_q", "P_bg767", "P_bg768", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn1.to_v", "P_bg767", "P_bg768", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn2.to_k", "P_bg769", "P_bg770", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn2.to_out.0", "P_bg771", "P_bg772", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn2.to_q", "P_bg773", "P_bg774", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.attn2.to_v", "P_bg775", "P_bg776", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.ff.net.0.proj", "P_bg777", "P_bg778", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.ff.net.2", "P_bg779", "P_bg780", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.norm1", "P_bg781"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.norm2", "P_bg781"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.3.norm3", "P_bg781"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn1.to_k", "P_bg782", "P_bg783", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn1.to_out.0", "P_bg782", "P_bg783", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn1.to_q", "P_bg782", "P_bg783", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn1.to_v", "P_bg782", "P_bg783", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn2.to_k", "P_bg784", "P_bg785", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn2.to_out.0", "P_bg786", "P_bg787", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn2.to_q", "P_bg788", "P_bg789", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.attn2.to_v", "P_bg790", "P_bg791", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.ff.net.0.proj", "P_bg792", "P_bg793", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.ff.net.2", "P_bg794", "P_bg795", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.norm1", "P_bg796"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.norm2", "P_bg796"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.4.norm3", "P_bg796"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn1.to_k", "P_bg797", "P_bg798", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn1.to_out.0", "P_bg797", "P_bg798", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn1.to_q", "P_bg797", "P_bg798", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn1.to_v", "P_bg797", "P_bg798", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn2.to_k", "P_bg799", "P_bg800", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn2.to_out.0", "P_bg801", "P_bg802", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn2.to_q", "P_bg803", "P_bg804", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.attn2.to_v", "P_bg805", "P_bg806", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.ff.net.0.proj", "P_bg807", "P_bg808", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.ff.net.2", "P_bg809", "P_bg810", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.norm1", "P_bg811"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.norm2", "P_bg811"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.5.norm3", "P_bg811"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn1.to_k", "P_bg812", "P_bg813", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn1.to_out.0", "P_bg812", "P_bg813", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn1.to_q", "P_bg812", "P_bg813", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn1.to_v", "P_bg812", "P_bg813", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn2.to_k", "P_bg814", "P_bg815", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn2.to_out.0", "P_bg816", "P_bg817", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn2.to_q", "P_bg818", "P_bg819", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.attn2.to_v", "P_bg820", "P_bg821", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.ff.net.0.proj", "P_bg822", "P_bg823", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.ff.net.2", "P_bg824", "P_bg825", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.norm1", "P_bg826"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.norm2", "P_bg826"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.6.norm3", "P_bg826"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn1.to_k", "P_bg827", "P_bg828", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn1.to_out.0", "P_bg827", "P_bg828", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn1.to_q", "P_bg827", "P_bg828", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn1.to_v", "P_bg827", "P_bg828", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn2.to_k", "P_bg829", "P_bg830", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn2.to_out.0", "P_bg831", "P_bg832", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn2.to_q", "P_bg833", "P_bg834", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.attn2.to_v", "P_bg835", "P_bg836", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.ff.net.0.proj", "P_bg837", "P_bg838", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.ff.net.2", "P_bg839", "P_bg840", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.norm1", "P_bg841"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.norm2", "P_bg841"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.7.norm3", "P_bg841"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn1.to_k", "P_bg842", "P_bg843", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn1.to_out.0", "P_bg842", "P_bg843", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn1.to_q", "P_bg842", "P_bg843", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn1.to_v", "P_bg842", "P_bg843", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn2.to_k", "P_bg844", "P_bg845", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn2.to_out.0", "P_bg846", "P_bg847", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn2.to_q", "P_bg848", "P_bg849", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.attn2.to_v", "P_bg850", "P_bg851", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.ff.net.0.proj", "P_bg852", "P_bg853", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.ff.net.2", "P_bg854", "P_bg855", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.norm1", "P_bg856"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.norm2", "P_bg856"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.8.norm3", "P_bg856"),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn1.to_k", "P_bg857", "P_bg858", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn1.to_out.0", "P_bg857", "P_bg858", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn1.to_q", "P_bg857", "P_bg858", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn1.to_v", "P_bg857", "P_bg858", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn2.to_k", "P_bg859", "P_bg860", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn2.to_out.0", "P_bg861", "P_bg862", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn2.to_q", "P_bg863", "P_bg864", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.attn2.to_v", "P_bg865", "P_bg866", bias=False),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.ff.net.0.proj", "P_bg867", "P_bg868", bias=True),
+            **dense("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.ff.net.2", "P_bg869", "P_bg870", bias=True),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.norm1", "P_bg871"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.norm2", "P_bg871"),
+            **norm("model.diffusion_model.output_blocks.1.1.transformer_blocks.9.norm3", "P_bg871"),
+            **easyblock("model.diffusion_model.output_blocks.2.0", "P_bg872", "P_bg873"),
+            **conv("model.diffusion_model.output_blocks.2.0.skip_connection", "P_bg874", "P_bg875"),
+            **norm("model.diffusion_model.output_blocks.2.1.norm", "P_bg876"),
+            **conv("model.diffusion_model.output_blocks.2.1.proj_in", "P_bg876", "P_bg877"),
+            **conv("model.diffusion_model.output_blocks.2.1.proj_out", "P_bg878", "P_bg879"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn1.to_k", "P_bg880", "P_bg881", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn1.to_out.0", "P_bg880", "P_bg881", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn1.to_q", "P_bg880", "P_bg881", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn1.to_v", "P_bg880", "P_bg881", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn2.to_k", "P_bg882", "P_bg883", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn2.to_out.0", "P_bg884", "P_bg885", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn2.to_q", "P_bg886", "P_bg887", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.attn2.to_v", "P_bg888", "P_bg889", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.ff.net.0.proj", "P_bg890", "P_bg891", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.ff.net.2", "P_bg892", "P_bg893", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.norm1", "P_bg894"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.norm2", "P_bg894"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.0.norm3", "P_bg894"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn1.to_k", "P_bg895", "P_bg896", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn1.to_out.0", "P_bg895", "P_bg896", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn1.to_q", "P_bg895", "P_bg896", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn1.to_v", "P_bg895", "P_bg896", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn2.to_k", "P_bg897", "P_bg898", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn2.to_out.0", "P_bg899", "P_bg900", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn2.to_q", "P_bg901", "P_bg902", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.attn2.to_v", "P_bg903", "P_bg904", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.ff.net.0.proj", "P_bg905", "P_bg906", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.ff.net.2", "P_bg907", "P_bg908", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.norm1", "P_bg909"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.norm2", "P_bg909"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.1.norm3", "P_bg909"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn1.to_k", "P_bg910", "P_bg911", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn1.to_out.0", "P_bg910", "P_bg911", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn1.to_q", "P_bg910", "P_bg911", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn1.to_v", "P_bg910", "P_bg911", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn2.to_k", "P_bg912", "P_bg913", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn2.to_out.0", "P_bg914", "P_bg915", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn2.to_q", "P_bg916", "P_bg917", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.attn2.to_v", "P_bg918", "P_bg919", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.ff.net.0.proj", "P_bg920", "P_bg921", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.ff.net.2", "P_bg922", "P_bg923", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.norm1", "P_bg924"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.norm2", "P_bg924"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.2.norm3", "P_bg924"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn1.to_k", "P_bg925", "P_bg926", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn1.to_out.0", "P_bg925", "P_bg926", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn1.to_q", "P_bg925", "P_bg926", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn1.to_v", "P_bg925", "P_bg926", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn2.to_k", "P_bg927", "P_bg928", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn2.to_out.0", "P_bg929", "P_bg930", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn2.to_q", "P_bg931", "P_bg932", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.attn2.to_v", "P_bg933", "P_bg934", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.ff.net.0.proj", "P_bg935", "P_bg936", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.ff.net.2", "P_bg937", "P_bg938", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.norm1", "P_bg939"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.norm2", "P_bg939"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.3.norm3", "P_bg939"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn1.to_k", "P_bg940", "P_bg941", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn1.to_out.0", "P_bg940", "P_bg941", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn1.to_q", "P_bg940", "P_bg941", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn1.to_v", "P_bg940", "P_bg941", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn2.to_k", "P_bg942", "P_bg943", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn2.to_out.0", "P_bg944", "P_bg945", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn2.to_q", "P_bg946", "P_bg947", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.attn2.to_v", "P_bg948", "P_bg949", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.ff.net.0.proj", "P_bg950", "P_bg951", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.ff.net.2", "P_bg952", "P_bg953", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.norm1", "P_bg954"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.norm2", "P_bg954"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.4.norm3", "P_bg954"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn1.to_k", "P_bg955", "P_bg956", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn1.to_out.0", "P_bg955", "P_bg956", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn1.to_q", "P_bg955", "P_bg956", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn1.to_v", "P_bg955", "P_bg956", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn2.to_k", "P_bg957", "P_bg958", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn2.to_out.0", "P_bg959", "P_bg960", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn2.to_q", "P_bg961", "P_bg962", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.attn2.to_v", "P_bg963", "P_bg964", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.ff.net.0.proj", "P_bg965", "P_bg966", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.ff.net.2", "P_bg967", "P_bg968", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.norm1", "P_bg969"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.norm2", "P_bg969"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.5.norm3", "P_bg969"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn1.to_k", "P_bg970", "P_bg971", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn1.to_out.0", "P_bg970", "P_bg971", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn1.to_q", "P_bg970", "P_bg971", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn1.to_v", "P_bg970", "P_bg971", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn2.to_k", "P_bg972", "P_bg973", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn2.to_out.0", "P_bg974", "P_bg975", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn2.to_q", "P_bg976", "P_bg977", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.attn2.to_v", "P_bg978", "P_bg979", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.ff.net.0.proj", "P_bg980", "P_bg981", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.ff.net.2", "P_bg982", "P_bg983", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.norm1", "P_bg984"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.norm2", "P_bg984"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.6.norm3", "P_bg984"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn1.to_k", "P_bg985", "P_bg986", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn1.to_out.0", "P_bg985", "P_bg986", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn1.to_q", "P_bg985", "P_bg986", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn1.to_v", "P_bg985", "P_bg986", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn2.to_k", "P_bg987", "P_bg988", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn2.to_out.0", "P_bg989", "P_bg990", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn2.to_q", "P_bg991", "P_bg992", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.attn2.to_v", "P_bg993", "P_bg994", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.ff.net.0.proj", "P_bg995", "P_bg996", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.ff.net.2", "P_bg997", "P_bg998", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.norm1", "P_bg999"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.norm2", "P_bg999"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.7.norm3", "P_bg999"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn1.to_k", "P_bg1000", "P_bg1001", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn1.to_out.0", "P_bg1000", "P_bg1001", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn1.to_q", "P_bg1000", "P_bg1001", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn1.to_v", "P_bg1000", "P_bg1001", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn2.to_k", "P_bg1002", "P_bg1003", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn2.to_out.0", "P_bg1004", "P_bg1005", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn2.to_q", "P_bg1006", "P_bg1007", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.attn2.to_v", "P_bg1008", "P_bg1009", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.ff.net.0.proj", "P_bg1010", "P_bg1011", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.ff.net.2", "P_bg1012", "P_bg1013", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.norm1", "P_bg1014"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.norm2", "P_bg1014"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.8.norm3", "P_bg1014"),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn1.to_k", "P_bg1015", "P_bg1016", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn1.to_out.0", "P_bg1015", "P_bg1016", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn1.to_q", "P_bg1015", "P_bg1016", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn1.to_v", "P_bg1015", "P_bg1016", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn2.to_k", "P_bg1017", "P_bg1018", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn2.to_out.0", "P_bg1019", "P_bg1020", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn2.to_q", "P_bg1021", "P_bg1022", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.attn2.to_v", "P_bg1023", "P_bg1024", bias=False),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.ff.net.0.proj", "P_bg1025", "P_bg1026", bias=True),
+            **dense("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.ff.net.2", "P_bg1027", "P_bg1028", bias=True),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.norm1", "P_bg1029"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.norm2", "P_bg1029"),
+            **norm("model.diffusion_model.output_blocks.2.1.transformer_blocks.9.norm3", "P_bg1029"),
+            **easyblock("model.diffusion_model.output_blocks.3.0", "P_bg1030", "P_bg1031"),
+            **conv("model.diffusion_model.output_blocks.3.0.skip_connection", "P_bg1032", "P_bg1033"),
+            **norm("model.diffusion_model.output_blocks.3.1.norm", "P_bg1034"),
+            **conv("model.diffusion_model.output_blocks.3.1.proj_in", "P_bg1034", "P_bg1035"),
+            **conv("model.diffusion_model.output_blocks.3.1.proj_out", "P_bg1036", "P_bg1037"),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn1.to_k", "P_bg1038", "P_bg1039", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn1.to_out.0", "P_bg1038", "P_bg1039", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn1.to_q", "P_bg1038", "P_bg1039", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn1.to_v", "P_bg1038", "P_bg1039", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn2.to_k", "P_bg1040", "P_bg1041", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn2.to_out.0", "P_bg1042", "P_bg1043", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn2.to_q", "P_bg1044", "P_bg1045", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.attn2.to_v", "P_bg1046", "P_bg1047", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.ff.net.0.proj", "P_bg1048", "P_bg1049", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.ff.net.2", "P_bg1050", "P_bg1051", bias=True),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.norm1", "P_bg1052"),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.norm2", "P_bg1052"),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.0.norm3", "P_bg1052"),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn1.to_k", "P_bg1053", "P_bg1054", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn1.to_out.0", "P_bg1053", "P_bg1054", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn1.to_q", "P_bg1053", "P_bg1054", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn1.to_v", "P_bg1053", "P_bg1054", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn2.to_k", "P_bg1055", "P_bg1056", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn2.to_out.0", "P_bg1057", "P_bg1058", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn2.to_q", "P_bg1059", "P_bg1060", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.attn2.to_v", "P_bg1061", "P_bg1062", bias=False),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.ff.net.0.proj", "P_bg1063", "P_bg1064", bias=True),
+            **dense("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.ff.net.2", "P_bg1065", "P_bg1066", bias=True),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.norm1", "P_bg1067"),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.norm2", "P_bg1067"),
+            **norm("model.diffusion_model.output_blocks.3.1.transformer_blocks.1.norm3", "P_bg1067"),
+            **easyblock("model.diffusion_model.output_blocks.4.0", "P_bg1068", "P_bg1069"),
+            **conv("model.diffusion_model.output_blocks.4.0.skip_connection", "P_bg1070", "P_bg1071"),
+            **norm("model.diffusion_model.output_blocks.4.1.norm", "P_bg1072"),
+            **conv("model.diffusion_model.output_blocks.4.1.proj_in", "P_bg1072", "P_bg1073"),
+            **conv("model.diffusion_model.output_blocks.4.1.proj_out", "P_bg1074", "P_bg1075"),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn1.to_k", "P_bg1076", "P_bg1077", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn1.to_out.0", "P_bg1076", "P_bg1077", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn1.to_q", "P_bg1076", "P_bg1077", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn1.to_v", "P_bg1076", "P_bg1077", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn2.to_k", "P_bg1078", "P_bg1079", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn2.to_out.0", "P_bg1080", "P_bg1081", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn2.to_q", "P_bg1082", "P_bg1083", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.attn2.to_v", "P_bg1084", "P_bg1085", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.ff.net.0.proj", "P_bg1086", "P_bg1087", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.ff.net.2", "P_bg1088", "P_bg1089", bias=True),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.norm1", "P_bg1090"),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.norm2", "P_bg1090"),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.0.norm3", "P_bg1090"),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn1.to_k", "P_bg1091", "P_bg1092", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn1.to_out.0", "P_bg1091", "P_bg1092", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn1.to_q", "P_bg1091", "P_bg1092", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn1.to_v", "P_bg1091", "P_bg1092", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn2.to_k", "P_bg1093", "P_bg1094", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn2.to_out.0", "P_bg1095", "P_bg1096", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn2.to_q", "P_bg1097", "P_bg1098", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.attn2.to_v", "P_bg1099", "P_bg1100", bias=False),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.ff.net.0.proj", "P_bg1101", "P_bg1102", bias=True),
+            **dense("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.ff.net.2", "P_bg1103", "P_bg1104", bias=True),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.norm1", "P_bg1105"),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.norm2", "P_bg1105"),
+            **norm("model.diffusion_model.output_blocks.4.1.transformer_blocks.1.norm3", "P_bg1105"),
+            **easyblock("model.diffusion_model.output_blocks.5.0", "P_bg1106", "P_bg1107"),
+            **conv("model.diffusion_model.output_blocks.5.0.skip_connection", "P_bg1108", "P_bg1109"),
+            **norm("model.diffusion_model.output_blocks.5.1.norm", "P_bg1110"),
+            **conv("model.diffusion_model.output_blocks.5.1.proj_in", "P_bg1110", "P_bg1111"),
+            **conv("model.diffusion_model.output_blocks.5.1.proj_out", "P_bg1112", "P_bg1113"),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn1.to_k", "P_bg1114", "P_bg1115", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn1.to_out.0", "P_bg1114", "P_bg1115", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn1.to_q", "P_bg1114", "P_bg1115", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn1.to_v", "P_bg1114", "P_bg1115", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn2.to_k", "P_bg1116", "P_bg1117", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn2.to_out.0", "P_bg1118", "P_bg1119", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn2.to_q", "P_bg1120", "P_bg1121", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.attn2.to_v", "P_bg1122", "P_bg1123", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.ff.net.0.proj", "P_bg1124", "P_bg1125", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.ff.net.2", "P_bg1126", "P_bg1127", bias=True),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.norm1", "P_bg1128"),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.norm2", "P_bg1128"),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.0.norm3", "P_bg1128"),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn1.to_k", "P_bg1129", "P_bg1130", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn1.to_out.0", "P_bg1129", "P_bg1130", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn1.to_q", "P_bg1129", "P_bg1130", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn1.to_v", "P_bg1129", "P_bg1130", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn2.to_k", "P_bg1131", "P_bg1132", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn2.to_out.0", "P_bg1133", "P_bg1134", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn2.to_q", "P_bg1135", "P_bg1136", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.attn2.to_v", "P_bg1137", "P_bg1138", bias=False),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.ff.net.0.proj", "P_bg1139", "P_bg1140", bias=True),
+            **dense("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.ff.net.2", "P_bg1141", "P_bg1142", bias=True),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.norm1", "P_bg1143"),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.norm2", "P_bg1143"),
+            **norm("model.diffusion_model.output_blocks.5.1.transformer_blocks.1.norm3", "P_bg1143"),
+            **conv("model.diffusion_model.output_blocks.5.2.conv", "P_bg1144", "P_bg1145"),
+            **easyblock("model.diffusion_model.output_blocks.6.0", "P_bg1146", "P_bg1147"),
+            **conv("model.diffusion_model.output_blocks.6.0.skip_connection", "P_bg1148", "P_bg1149"),
+            **easyblock("model.diffusion_model.output_blocks.7.0", "P_bg1150", "P_bg1151"),
+            **conv("model.diffusion_model.output_blocks.7.0.skip_connection", "P_bg1152", "P_bg1153"),
+            **easyblock("model.diffusion_model.output_blocks.8.0", "P_bg1154", "P_bg1155"),
+            **conv("model.diffusion_model.output_blocks.8.0.skip_connection", "P_bg1156", "P_bg1157"),
+            #Text Encoder 1
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.position_embedding", "P_bg1158", "P_bg1159", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.token_embedding", "P_bg1160", "P_bg1161", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc1", "P_bg1164", "P_bg1165", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc2", "P_bg1166", "P_bg1167", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.k_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.out_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.q_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.v_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc1", "P_bg1172", "P_bg1173", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc2", "P_bg1174", "P_bg1175", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.k_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.out_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.q_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.v_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc1", "P_bg1180", "P_bg1181", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc2", "P_bg1182", "P_bg1183", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.k_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.out_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.q_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.v_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc1", "P_bg1188", "P_bg1189", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc2", "P_bg1190", "P_bg1191", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.k_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.out_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.q_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.v_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc1", "P_bg1196", "P_bg1197", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc2", "P_bg1198", "P_bg1199", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.k_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.out_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.q_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.v_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc1", "P_bg1204", "P_bg1205", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc2", "P_bg1206", "P_bg1207", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.k_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.out_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.q_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.v_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc1", "P_bg1212", "P_bg1213", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc2", "P_bg1214", "P_bg1215", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.k_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.out_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.q_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.v_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc1", "P_bg1220", "P_bg1221", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc2", "P_bg1222", "P_bg1223", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.k_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.out_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.q_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.v_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc1", "P_bg1228", "P_bg1229", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc2", "P_bg1230", "P_bg1231", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.k_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.out_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.q_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.v_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc1", "P_bg1236", "P_bg1237", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc2", "P_bg1238", "P_bg1239", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.k_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.out_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.q_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.v_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc1", "P_bg1244", "P_bg1245", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc2", "P_bg1246", "P_bg1247", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.k_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.out_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.q_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.v_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc1", "P_bg1252", "P_bg1253", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc2", "P_bg1254", "P_bg1255", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.k_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.out_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.q_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.v_proj", "P_bg1256", "P_bg1257", bias=False),
+            #Text Encoder 2
+            **dense("conditioner.embedders.1.model.token_embedding", "P_bg1260", "P_bg1261", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.in_proj_weight", "P_bg1262", "P_bg1263", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.out_proj", "P_bg1262", "P_bg1263", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_fc", "P_bg1266", "P_bg1267", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.in_proj_weight", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.out_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_fc", "P_bg1272", "P_bg1273", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.in_proj_weight", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.out_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_fc", "P_bg1278", "P_bg1279", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.in_proj_weight", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.out_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_fc", "P_bg1284", "P_bg1285", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.in_proj_weight", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.out_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_fc", "P_bg1290", "P_bg1291", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.in_proj_weight", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.out_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_fc", "P_bg1296", "P_bg1297", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.in_proj_weight", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.out_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_fc", "P_bg1302", "P_bg1303", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.in_proj_weight", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.out_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_fc", "P_bg1308", "P_bg1309", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.in_proj_weight", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.out_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_fc", "P_bg1314", "P_bg1315", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.in_proj_weight", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.out_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_fc", "P_bg1320", "P_bg1321", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.in_proj_weight", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.out_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_fc", "P_bg1326", "P_bg1327", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.in_proj_weight", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.out_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_fc", "P_bg1332", "P_bg1333", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.in_proj_weight", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.out_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_fc", "P_bg1338", "P_bg1339", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.in_proj_weight", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.out_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_fc", "P_bg1344", "P_bg1345", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.in_proj_weight", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.out_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_fc", "P_bg1350", "P_bg1351", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.in_proj_weight", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.out_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_fc", "P_bg1356", "P_bg1357", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.in_proj_weight", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.out_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_fc", "P_bg1362", "P_bg1363", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.in_proj_weight", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.out_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_fc", "P_bg1368", "P_bg1369", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.in_proj_weight", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.out_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_fc", "P_bg1374", "P_bg1375", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.in_proj_weight", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.out_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_fc", "P_bg1380", "P_bg1381", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.in_proj_weight", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.out_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_fc", "P_bg1386", "P_bg1387", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.in_proj_weight", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.out_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_fc", "P_bg1392", "P_bg1393", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.in_proj_weight", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.out_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_fc", "P_bg1398", "P_bg1399", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.in_proj_weight", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.out_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_fc", "P_bg1404", "P_bg1405", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.in_proj_weight", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.out_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_fc", "P_bg1410", "P_bg1411", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.in_proj_weight", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.out_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_fc", "P_bg1416", "P_bg1417", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.in_proj_weight", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.out_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_fc", "P_bg1422", "P_bg1423", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.in_proj_weight", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.out_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_fc", "P_bg1428", "P_bg1429", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.in_proj_weight", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.out_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_fc", "P_bg1434", "P_bg1435", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.in_proj_weight", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.out_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_fc", "P_bg1440", "P_bg1441", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.in_proj_weight", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.out_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_fc", "P_bg1446", "P_bg1447", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.in_proj_weight", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.out_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_fc", "P_bg1452", "P_bg1453", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_proj", "P_bg1454", "P_bg1455", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.position_embedding", "P_bg1158", "P_bg1159", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.token_embedding", "P_bg1160", "P_bg1161", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc1", "P_bg1164", "P_bg1165", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc2", "P_bg1166", "P_bg1167", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.k_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.out_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.q_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.v_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc1", "P_bg1172", "P_bg1173", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc2", "P_bg1174", "P_bg1175", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.k_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.out_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.q_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.v_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc1", "P_bg1180", "P_bg1181", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc2", "P_bg1182", "P_bg1183", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.k_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.out_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.q_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.v_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc1", "P_bg1188", "P_bg1189", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc2", "P_bg1190", "P_bg1191", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.k_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.out_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.q_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.v_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc1", "P_bg1196", "P_bg1197", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc2", "P_bg1198", "P_bg1199", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.k_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.out_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.q_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.v_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc1", "P_bg1204", "P_bg1205", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc2", "P_bg1206", "P_bg1207", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.k_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.out_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.q_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.v_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc1", "P_bg1212", "P_bg1213", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc2", "P_bg1214", "P_bg1215", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.k_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.out_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.q_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.v_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc1", "P_bg1220", "P_bg1221", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc2", "P_bg1222", "P_bg1223", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.k_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.out_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.q_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.v_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc1", "P_bg1228", "P_bg1229", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc2", "P_bg1230", "P_bg1231", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.k_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.out_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.q_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.v_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc1", "P_bg1236", "P_bg1237", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc2", "P_bg1238", "P_bg1239", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.k_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.out_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.q_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.v_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc1", "P_bg1244", "P_bg1245", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc2", "P_bg1246", "P_bg1247", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.k_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.out_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.q_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.v_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc1", "P_bg1252", "P_bg1253", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc2", "P_bg1254", "P_bg1255", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.k_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.out_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.q_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.v_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.1.model.token_embedding", "P_bg1260", "P_bg1261", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.in_proj_weight", "P_bg1262", "P_bg1263", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.out_proj", "P_bg1262", "P_bg1263", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_fc", "P_bg1266", "P_bg1267", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.in_proj_weight", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.out_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_fc", "P_bg1272", "P_bg1273", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.in_proj_weight", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.out_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_fc", "P_bg1278", "P_bg1279", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.in_proj_weight", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.out_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_fc", "P_bg1284", "P_bg1285", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.in_proj_weight", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.out_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_fc", "P_bg1290", "P_bg1291", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.in_proj_weight", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.out_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_fc", "P_bg1296", "P_bg1297", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.in_proj_weight", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.out_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_fc", "P_bg1302", "P_bg1303", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.in_proj_weight", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.out_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_fc", "P_bg1308", "P_bg1309", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.in_proj_weight", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.out_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_fc", "P_bg1314", "P_bg1315", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.in_proj_weight", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.out_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_fc", "P_bg1320", "P_bg1321", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.in_proj_weight", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.out_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_fc", "P_bg1326", "P_bg1327", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.in_proj_weight", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.out_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_fc", "P_bg1332", "P_bg1333", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.in_proj_weight", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.out_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_fc", "P_bg1338", "P_bg1339", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.in_proj_weight", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.out_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_fc", "P_bg1344", "P_bg1345", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.in_proj_weight", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.out_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_fc", "P_bg1350", "P_bg1351", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.in_proj_weight", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.out_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_fc", "P_bg1356", "P_bg1357", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.in_proj_weight", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.out_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_fc", "P_bg1362", "P_bg1363", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.in_proj_weight", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.out_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_fc", "P_bg1368", "P_bg1369", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.in_proj_weight", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.out_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_fc", "P_bg1374", "P_bg1375", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.in_proj_weight", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.out_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_fc", "P_bg1380", "P_bg1381", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.in_proj_weight", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.out_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_fc", "P_bg1386", "P_bg1387", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.in_proj_weight", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.out_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_fc", "P_bg1392", "P_bg1393", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.in_proj_weight", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.out_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_fc", "P_bg1398", "P_bg1399", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.in_proj_weight", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.out_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_fc", "P_bg1404", "P_bg1405", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.in_proj_weight", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.out_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_fc", "P_bg1410", "P_bg1411", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.in_proj_weight", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.out_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_fc", "P_bg1416", "P_bg1417", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.in_proj_weight", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.out_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_fc", "P_bg1422", "P_bg1423", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.in_proj_weight", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.out_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_fc", "P_bg1428", "P_bg1429", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.in_proj_weight", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.out_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_fc", "P_bg1434", "P_bg1435", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.in_proj_weight", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.out_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_fc", "P_bg1440", "P_bg1441", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.in_proj_weight", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.out_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_fc", "P_bg1446", "P_bg1447", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.in_proj_weight", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.out_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_fc", "P_bg1452", "P_bg1453", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_proj", "P_bg1454", "P_bg1455", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.position_embedding", "P_bg1158", "P_bg1159", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.token_embedding", "P_bg1160", "P_bg1161", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc1", "P_bg1164", "P_bg1165", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc2", "P_bg1166", "P_bg1167", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.k_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.out_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.q_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.v_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc1", "P_bg1172", "P_bg1173", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc2", "P_bg1174", "P_bg1175", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.k_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.out_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.q_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.v_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc1", "P_bg1180", "P_bg1181", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc2", "P_bg1182", "P_bg1183", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.k_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.out_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.q_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.v_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc1", "P_bg1188", "P_bg1189", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc2", "P_bg1190", "P_bg1191", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.k_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.out_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.q_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.v_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc1", "P_bg1196", "P_bg1197", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc2", "P_bg1198", "P_bg1199", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.k_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.out_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.q_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.v_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc1", "P_bg1204", "P_bg1205", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc2", "P_bg1206", "P_bg1207", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.k_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.out_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.q_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.v_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc1", "P_bg1212", "P_bg1213", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc2", "P_bg1214", "P_bg1215", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.k_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.out_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.q_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.v_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc1", "P_bg1220", "P_bg1221", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc2", "P_bg1222", "P_bg1223", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.k_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.out_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.q_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.v_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc1", "P_bg1228", "P_bg1229", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc2", "P_bg1230", "P_bg1231", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.k_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.out_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.q_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.v_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc1", "P_bg1236", "P_bg1237", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc2", "P_bg1238", "P_bg1239", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.k_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.out_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.q_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.v_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc1", "P_bg1244", "P_bg1245", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc2", "P_bg1246", "P_bg1247", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.k_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.out_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.q_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.v_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc1", "P_bg1252", "P_bg1253", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc2", "P_bg1254", "P_bg1255", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.k_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.out_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.q_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.v_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.1.model.token_embedding", "P_bg1260", "P_bg1261", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.in_proj_weight", "P_bg1262", "P_bg1263", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.out_proj", "P_bg1262", "P_bg1263", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_fc", "P_bg1266", "P_bg1267", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.in_proj_weight", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.out_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_fc", "P_bg1272", "P_bg1273", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.in_proj_weight", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.out_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_fc", "P_bg1278", "P_bg1279", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.in_proj_weight", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.out_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_fc", "P_bg1284", "P_bg1285", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.in_proj_weight", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.out_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_fc", "P_bg1290", "P_bg1291", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.in_proj_weight", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.out_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_fc", "P_bg1296", "P_bg1297", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.in_proj_weight", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.out_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_fc", "P_bg1302", "P_bg1303", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.in_proj_weight", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.out_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_fc", "P_bg1308", "P_bg1309", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.in_proj_weight", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.out_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_fc", "P_bg1314", "P_bg1315", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.in_proj_weight", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.out_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_fc", "P_bg1320", "P_bg1321", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.in_proj_weight", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.out_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_fc", "P_bg1326", "P_bg1327", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.in_proj_weight", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.out_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_fc", "P_bg1332", "P_bg1333", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.in_proj_weight", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.out_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_fc", "P_bg1338", "P_bg1339", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.in_proj_weight", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.out_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_fc", "P_bg1344", "P_bg1345", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.in_proj_weight", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.out_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_fc", "P_bg1350", "P_bg1351", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.in_proj_weight", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.out_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_fc", "P_bg1356", "P_bg1357", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.in_proj_weight", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.out_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_fc", "P_bg1362", "P_bg1363", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.in_proj_weight", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.out_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_fc", "P_bg1368", "P_bg1369", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.in_proj_weight", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.out_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_fc", "P_bg1374", "P_bg1375", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.in_proj_weight", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.out_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_fc", "P_bg1380", "P_bg1381", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.in_proj_weight", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.out_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_fc", "P_bg1386", "P_bg1387", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.in_proj_weight", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.out_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_fc", "P_bg1392", "P_bg1393", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.in_proj_weight", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.out_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_fc", "P_bg1398", "P_bg1399", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.in_proj_weight", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.out_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_fc", "P_bg1404", "P_bg1405", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.in_proj_weight", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.out_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_fc", "P_bg1410", "P_bg1411", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.in_proj_weight", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.out_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_fc", "P_bg1416", "P_bg1417", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.in_proj_weight", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.out_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_fc", "P_bg1422", "P_bg1423", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.in_proj_weight", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.out_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_fc", "P_bg1428", "P_bg1429", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.in_proj_weight", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.out_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_fc", "P_bg1434", "P_bg1435", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.in_proj_weight", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.out_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_fc", "P_bg1440", "P_bg1441", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.in_proj_weight", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.out_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_fc", "P_bg1446", "P_bg1447", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.in_proj_weight", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.out_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_fc", "P_bg1452", "P_bg1453", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_proj", "P_bg1454", "P_bg1455", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.position_embedding", "P_bg1158", "P_bg1159", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.embeddings.token_embedding", "P_bg1160", "P_bg1161", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc1", "P_bg1164", "P_bg1165", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.mlp.fc2", "P_bg1166", "P_bg1167", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.k_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.out_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.q_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.0.self_attn.v_proj", "P_bg1168", "P_bg1169", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc1", "P_bg1172", "P_bg1173", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.mlp.fc2", "P_bg1174", "P_bg1175", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.k_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.out_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.q_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.1.self_attn.v_proj", "P_bg1176", "P_bg1177", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc1", "P_bg1180", "P_bg1181", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.mlp.fc2", "P_bg1182", "P_bg1183", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.k_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.out_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.q_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.10.self_attn.v_proj", "P_bg1184", "P_bg1185", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc1", "P_bg1188", "P_bg1189", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.mlp.fc2", "P_bg1190", "P_bg1191", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.k_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.out_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.q_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.11.self_attn.v_proj", "P_bg1192", "P_bg1193", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc1", "P_bg1196", "P_bg1197", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.mlp.fc2", "P_bg1198", "P_bg1199", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.k_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.out_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.q_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.2.self_attn.v_proj", "P_bg1200", "P_bg1201", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc1", "P_bg1204", "P_bg1205", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.mlp.fc2", "P_bg1206", "P_bg1207", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.k_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.out_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.q_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.3.self_attn.v_proj", "P_bg1208", "P_bg1209", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc1", "P_bg1212", "P_bg1213", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.mlp.fc2", "P_bg1214", "P_bg1215", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.k_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.out_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.q_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.4.self_attn.v_proj", "P_bg1216", "P_bg1217", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc1", "P_bg1220", "P_bg1221", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.mlp.fc2", "P_bg1222", "P_bg1223", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.k_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.out_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.q_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.5.self_attn.v_proj", "P_bg1224", "P_bg1225", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc1", "P_bg1228", "P_bg1229", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.mlp.fc2", "P_bg1230", "P_bg1231", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.k_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.out_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.q_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.6.self_attn.v_proj", "P_bg1232", "P_bg1233", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc1", "P_bg1236", "P_bg1237", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.mlp.fc2", "P_bg1238", "P_bg1239", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.k_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.out_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.q_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.7.self_attn.v_proj", "P_bg1240", "P_bg1241", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc1", "P_bg1244", "P_bg1245", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.mlp.fc2", "P_bg1246", "P_bg1247", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.k_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.out_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.q_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.8.self_attn.v_proj", "P_bg1248", "P_bg1249", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc1", "P_bg1252", "P_bg1253", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.mlp.fc2", "P_bg1254", "P_bg1255", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.k_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.out_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.q_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.0.transformer.text_model.encoder.layers.9.self_attn.v_proj", "P_bg1256", "P_bg1257", bias=False),
+            **dense("conditioner.embedders.1.model.token_embedding", "P_bg1260", "P_bg1261", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.in_proj_weight", "P_bg1262", "P_bg1263", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.attn.out_proj", "P_bg1262", "P_bg1263", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_fc", "P_bg1266", "P_bg1267", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.0.mlp.c_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.in_proj_weight", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.attn.out_proj", "P_bg1268", "P_bg1269", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_fc", "P_bg1272", "P_bg1273", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.1.mlp.c_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.in_proj_weight", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.attn.out_proj", "P_bg1274", "P_bg1275", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_fc", "P_bg1278", "P_bg1279", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.10.mlp.c_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.in_proj_weight", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.attn.out_proj", "P_bg1280", "P_bg1281", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_fc", "P_bg1284", "P_bg1285", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.11.mlp.c_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.in_proj_weight", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.attn.out_proj", "P_bg1286", "P_bg1287", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_fc", "P_bg1290", "P_bg1291", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.12.mlp.c_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.in_proj_weight", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.attn.out_proj", "P_bg1292", "P_bg1293", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_fc", "P_bg1296", "P_bg1297", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.13.mlp.c_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.in_proj_weight", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.attn.out_proj", "P_bg1298", "P_bg1299", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_fc", "P_bg1302", "P_bg1303", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.14.mlp.c_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.in_proj_weight", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.attn.out_proj", "P_bg1304", "P_bg1305", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_fc", "P_bg1308", "P_bg1309", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.15.mlp.c_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.in_proj_weight", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.attn.out_proj", "P_bg1310", "P_bg1311", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_fc", "P_bg1314", "P_bg1315", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.16.mlp.c_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.in_proj_weight", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.attn.out_proj", "P_bg1316", "P_bg1317", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_fc", "P_bg1320", "P_bg1321", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.17.mlp.c_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.in_proj_weight", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.attn.out_proj", "P_bg1322", "P_bg1323", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_fc", "P_bg1326", "P_bg1327", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.18.mlp.c_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.in_proj_weight", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.attn.out_proj", "P_bg1328", "P_bg1329", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_fc", "P_bg1332", "P_bg1333", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.19.mlp.c_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.in_proj_weight", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.attn.out_proj", "P_bg1334", "P_bg1335", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_fc", "P_bg1338", "P_bg1339", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.2.mlp.c_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.in_proj_weight", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.attn.out_proj", "P_bg1340", "P_bg1341", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_fc", "P_bg1344", "P_bg1345", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.20.mlp.c_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.in_proj_weight", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.attn.out_proj", "P_bg1346", "P_bg1347", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_fc", "P_bg1350", "P_bg1351", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.21.mlp.c_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.in_proj_weight", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.attn.out_proj", "P_bg1352", "P_bg1353", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_fc", "P_bg1356", "P_bg1357", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.22.mlp.c_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.in_proj_weight", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.attn.out_proj", "P_bg1358", "P_bg1359", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_fc", "P_bg1362", "P_bg1363", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.23.mlp.c_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.in_proj_weight", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.attn.out_proj", "P_bg1364", "P_bg1365", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_fc", "P_bg1368", "P_bg1369", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.24.mlp.c_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.in_proj_weight", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.attn.out_proj", "P_bg1370", "P_bg1371", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_fc", "P_bg1374", "P_bg1375", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.25.mlp.c_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.in_proj_weight", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.attn.out_proj", "P_bg1376", "P_bg1377", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_fc", "P_bg1380", "P_bg1381", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.26.mlp.c_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.in_proj_weight", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.attn.out_proj", "P_bg1382", "P_bg1383", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_fc", "P_bg1386", "P_bg1387", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.27.mlp.c_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.in_proj_weight", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.attn.out_proj", "P_bg1388", "P_bg1389", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_fc", "P_bg1392", "P_bg1393", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.28.mlp.c_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.in_proj_weight", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.attn.out_proj", "P_bg1394", "P_bg1395", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_fc", "P_bg1398", "P_bg1399", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.29.mlp.c_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.in_proj_weight", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.attn.out_proj", "P_bg1400", "P_bg1401", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_fc", "P_bg1404", "P_bg1405", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.3.mlp.c_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.in_proj_weight", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.attn.out_proj", "P_bg1406", "P_bg1407", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_fc", "P_bg1410", "P_bg1411", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.30.mlp.c_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.in_proj_weight", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.attn.out_proj", "P_bg1412", "P_bg1413", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_fc", "P_bg1416", "P_bg1417", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.31.mlp.c_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.in_proj_weight", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.attn.out_proj", "P_bg1418", "P_bg1419", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_fc", "P_bg1422", "P_bg1423", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.4.mlp.c_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.in_proj_weight", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.attn.out_proj", "P_bg1424", "P_bg1425", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_fc", "P_bg1428", "P_bg1429", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.5.mlp.c_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.in_proj_weight", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.attn.out_proj", "P_bg1430", "P_bg1431", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_fc", "P_bg1434", "P_bg1435", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.6.mlp.c_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.in_proj_weight", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.attn.out_proj", "P_bg1436", "P_bg1437", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_fc", "P_bg1440", "P_bg1441", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.7.mlp.c_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.in_proj_weight", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.attn.out_proj", "P_bg1442", "P_bg1443", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_fc", "P_bg1446", "P_bg1447", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.8.mlp.c_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.in_proj_weight", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.attn.out_proj", "P_bg1448", "P_bg1449", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_fc", "P_bg1452", "P_bg1453", bias=False),
+            **dense("conditioner.embedders.1.model.transformer.resblocks.9.mlp.c_proj", "P_bg1454", "P_bg1455", bias=False),
+        }
+    )
